ISSUE Currently ForwardingRulesManager installs the proactive mode flows through the static flow configuration path. This allows these controller generated flows to be listed in the flow programmer UI. The drawback of this approach is that now the proactive flows are saved to FRM startup configuration, when user saves the config. So when FRM boots up, it will program the proactive flows on the switch whenever the switch connects to the controller, regardless of the actual switch forwarding mode configured on switch mgr.
This is not correct as the notion of proactive forwarding mode associated to a
switch is controlled by switch manager, through sm UI. The trigger to install the
proactive flows should come from switch manager, FRM should not save these internal
generated flows to its startup config, nor program them on its own.

CHANGES
1) On FRM config save, the internal generated static flows will not be added to the
list that will be actually saved.
2) In Switch Mgr, when the switch connects, while reading the respective switch
configuration (if any) the mode is learnt. If it is set to proactive, the mode change
event is being played.
3) When a switch disconnect, FRM has to remove the internal static flow configurations
for that switch.
4) Switch Mgr isSpecial method to correctly report special node conenctor (method should
be provided by NodeConnector class in future).
5) Adding the startup directory to the distribution to allow write/read of startup
configuration files.
6) Have FlowConfig class to expose a method to retrieve the configured forwarding mode,
so that internal coding is hidden to clients.

Signed-off-by: aboch <aboch@cisco.com>

Change-Id: Id628042b7d656d95ace87a1e72c13f2950e6795c
Signed-off-by: aboch <aboch@cisco.com>