Added support for RPCs to Binding-Aware SAL
Added support for RPCs to Binding-Aware SAL. The implementations
of the RPCs are mapped to interfaces describing RPC contract.

The registered implementation is registered for all RPCs described
in the interface and only one registered implementation of the
interface could exists at-the time.

It is possible to register implementation of derived interface,
but client must request the derived interface to access its
implementation of rpcs defined in super interface.

Signed-off-by: Tony Tkacik <ttkacik@cisco.com>