Update patch set 3
Patch Set 3:

In my environment, I tried deleting the org/opendaylight in my maven repository and do a mvn clean install, the integration test was able to find the maven bundles and passed successfully.  I'm not sure if it's the settings I have in the pom.xml or Jenkins settings that's causing the unstable build.

Patch-set: 3