Redirecting Caught and Uncaught Exceptions to OSGI Console and Log File
The existing mechanism int the Controller allows the exceptions to be printed only
the console. This applies to both caught and uncaught exception. If the console
buffer is not too large and gets overwritten or gets cleared, there is no way for
the user to know what exceptions occurred. This commit implements a new mechanism
by which both types of exceptions will get printed on the console as well as logged
to the file.

Signed-off-by: Maurice Qureshi <maquresh@cisco.com>