Update patch set 1
Patch Set 1:

This patch is still Work-in-Progress to showcase updates to the SAL and adding extensibility also to the southbound plugins (in this case Openflow plugin) - by providing
clearer abstractions do subclass define new Action
and also registering their Openflow counterpart to the
Openflow Plugin.

Feel free to review changes, suggest API changes and so on.

Please DO NOT MERGE right now.

Patch-set: 1