Add PDE OSGi launcher for the controller project
- Created a p2site distribution which will place in a p2 repository
all the artifacts that are part of the
opendaylight/commons/opendaylight/pom.xml parent pom. The p2site will
be built by the controller-merge job on jenkins and will be used as
update site for the "opendaylight" target platform.
- Created a target-platform to be used by PDE when used in eclipse
- Create an OSGi PDE launcher to be able to launch the controller
without the assembly phase to be required.
All this enhancements are to address some of the eclipse usability
issues as documented in:
https://bugs.opendaylight.org/show_bug.cgi?id=21

Change-Id: I559b92ff52011cd69d306b43edd2531901aac85f
Signed-off-by: Giovanni Meo <gmeo@cisco.com>