Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

(1 inline comment)

I tried out this patch and it did enable the creation of a p2 site. The included target was not settable because there was not a p2 site at the target location.  Also the included launch did not work because the projects in Eclipse are Maven based projects rather than PDE-based, so none of the OpenDaylight modules were found and made available for launch.  

I would be much more comfortable accepting a patch that includes a way to generate PDE meta-data for each bundle in the workspace, with a working target and a functioning launch.  If this is actually available somehow already please let me know and I can update this.

Patch-set: 1