Add PDE OSGi launcher for the controller project
- Created a p2site distribution which will place in a p2 repository
all the artifacts that are part of the
opendaylight/commons/opendaylight/pom.xml parent pom. The p2site will
be built by the controller-merge job on jenkins and will be used as
update site for the "opendaylight" target platform.
- Created a target-platform "opendaylight" to be used by PDE when used
in eclipse, this target platform is supposed to fetch the artifact
from a remote P2 site that will be hosted on jenkins after this patch
will get in.
- Create an OSGi PDE launcher "opendaylight-osgi-launcher" to be able
to launch the controller without the assembly phase to be required,
this one works with the target-platform "opendaylight"
- Created a target-platform "opendaylight-local" to be used by PDE
with the assumptions that the p2 site reside local and has been built
in the p2site distribution.
All this enhancements are to address some of the eclipse usability
issues as documented in:
https://bugs.opendaylight.org/show_bug.cgi?id=21

Change-Id: I559b92ff52011cd69d306b43edd2531901aac85f
Signed-off-by: Giovanni Meo <gmeo@cisco.com>