Update patch set 1
Patch Set 1:

Hi David,

Yes, Alessandro pointed out the white space issue and I've taken care of that in my next check-in. I referred to Ed's post to continue modifying a previous check-in: http://edwarnicke.org/2013/05/29/contributing-a-continuation-of-a-not-yet-merged-gerrit-change-in-opendaylight/ .
But I am missing something in the process, I've landed with gerrit 415 as the final check-in. I apologize for the multiple entries. I'll make sure it doesn't happen again.

--
Aditya

Patch-set: 1