Update patch set 1
Patch Set 1:

Hi Aditya-
Ya in general Ed's method applies to additional new code that depends on the unaccepted patch.  The vast majority of the time (and definitely in this case) you would want to continue ammending this commit until it meets the suggestions of the community.  By way of example, there is no easy way for me to verify that the spacing/tab problems introduced in this patch are for sure removed in 415.

Thanks!

Patch-set: 1