Update patch set 1
Patch Set 1:

Hi Alessandro,

For the record, I have created this patch with all the changes and zero dependency on any of the previous commits. This was after my conversation with Ed on the #opendaylight IRC channel. Whenever you merge, this is the only one that needs to be merged and abandon all the previous ones.

Thanks!
Aditya

Patch-set: 1