Update patch set 1
Patch Set 1:

Hi Tony-
Thumbs up on the versionAsInProject, I was wondering if that was possible.  To the second point in your commit message, did you miss checking in the corresponding POMs?

Also, it would probably still be useful to have ITs attached to the install goal, but would be useful to be able to disable only the ITs on the install phase, ie:

mvn install -DskipITs

Patch-set: 1