Checkstyle enforcer
Started to enforce checkstyle, if the code style doesn't match the
checkstyle settings build will break. This is needed in order to make
sure folks comply to one code style and don't start diverging.

For now the enforcement is only limited to:
- make sure no TABS are present in the java files
- make sure that no trailing whitespaces are present

Along with code style enforcement updates all the files that didn't
comply else compilation would not be successful.

Added new exclusion in .gitignore

Added new parent for commons.opendaylight, please read the README in
the commons.parent project for the reason for it.

Make sure the violation of checkstyle are reported on the console to
avoid for a developer to be able to dig through the log in the target
directory

Change-Id: I4e35aa74e685e5cebb8182caf0b3b39da0906948
Signed-off-by: Giovanni Meo <gmeo@cisco.com>