Checkstyle enforcer
Started to enforce checkstyle, if the code style doesn't match the
checkstyle settings build will break. This is needed in order to make
sure folks comply to one code style and don't start diverging.

For now the enforcement is only limited to:
- make sure no TABS are present in the java files
- make sure that no trailing whitespaces are present

Along with code style enforcement updates all the files that didn't
comply else compilation would not be successful.

Added new exclusion in .gitignore

Change-Id: Ic95da373575cec3cfe0c99b31fce92f66bc41263
Signed-off-by: Giovanni Meo <gmeo@cisco.com>