Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

(6 inline comments)

Do not catch exceptions and fail a test without them, e.g. JSONException is thrown if the field is missing, and you just fail the test without any explanation, so it is harder to debug. Change the method signature to throws JSONException and the exception will fail the test with more describing message.

Patch-set: 1