Update patch set 1
Patch Set 1:

It will throw "FileNotFoundException" as it requires the ".conf" files to run the test successfully.

The purpose of the test was totest ConfigurationService.saveConfiguration() which loads the .conf files for SwitchManager and save the configuration. So I need those .conf files in the startup directory to run the test successfully.

Patch-set: 1