Update patch set 3
Patch Set 3:

Kalvin, it would be great to have the patch not to depend on the Rob Adams one as it's currently, the reason is that gerrit will not merge this till the depending one is not merged. The Rob Adams patch cannot merge simply because of the fact that jenkins cannot verify it hence it's catch 22. So please submit a new patch not starting from a view with Rob Adams patch, this can be done by: 
1) creating a branch in git synched to current HEAD, lets call it "staging"
2) in staging area do a git pull --squash . <branch where there is code> 
3) Resolve conflicts and commit
4) push the update.

Patch-set: 3