Update patch set 2
Patch Set 2:

A (non complete) set of REST APIs for subnets is already present in subnets.northbound.

I am not clear why this commit is adding this a new full set to switchmanager.northbound instead of enhancing the existing one. Unless it was discussed to move subnets under switch mgr northbound which I was not aware of.

Patch-set: 2