Update patch set 5
Patch Set 5: Looks good to me, approved

Just for tracking, wanted to leave this comment.

Idea is that for each rest API (API3) there should be a correspondent API2 (service API) so that northbound or web functions would just make one call to the service and translate the returned Status to the correspondent Response.Status. This way all the validation is contained in the backend service that owns the configuration object.

I notice now that ISwitchManager does not provide such API, like modifySubnetConfig(SubnetConfig curent, SubnetConfig newone). We should probably add it in future.

Patch-set: 5