Create change
Patch-set: 1
Change-id: I0ffad694b1a21ff4ec7e2139bb81a398412b9de0
Subject: Connection Manager infrastructure supporting the Clustering Active-Active requirement. This commit brings in the following changes : 1. Connection Manager as a functional module that runs above SAL providing common services to various south-bound plugins.    The initial version of Connection Manager takes in Node as the only argument and let the Connection Management Scheme to determine the fate of the connection.    Primary use-case that is addressed here is the HA/Clustering's Active-Active mode with the ability for Nodes to connect to different controller in a cluster.
Branch: refs/heads/master
Commit: 59fdd1aa84189a4672326f225235f747f6eb9a66