Update patch set 1
Patch Set 1:

Yes, this patch was a case of trying to get something out first and something better out second.

It should give some flavor of what's coming, but I'm in the process of producing another patch set that is based on the current HEAD and has correct white space.

Patch-set: 1