StatisticsManager "started" method may fail
"started" method in the
org.opendaylight.controller.statisticsmanager.internal.StatisticsManager
occasionally can fail to complete, this is because the method tries to
"put" data in clustered caches without checking if the value being
inserted is null or no. Infinispan caches in fact don't support null
value insertion and hence an exception is thrown preventing subsequent
calls to continue. The fix is make sure we don't insert NULL values.

Change-Id: I0037931d73e7e7df9535ef83509e770b45e81c72
Signed-off-by: Giovanni Meo <gmeo@cisco.com>