Update patch set 2
Patch Set 2: Looks good to me, approved

I didnt see your comment back in Gerrit regarding the usage of coordinator vs isLocal. 
The idea of using Co-ordinator vs local controller to trigger the Routing update is debatable depending on the use-case and the condition that is being considered.  

The current diffs uses the coordinator approach for the benefits it provides. I am okay to approve this approach. But we can come back to it if practical reasoning forces us to look back at the approach of local controller being the master for the routing update to the applications above.

Patch-set: 2
Label: SUBM=+1