protocol.openflow_netty cleanup
- Avoid code duplication between protocol.openflow and
protocol.openflow_netty. The bundle protocol.openflow_netty can
largely be equated to a bundle that is sub-classing (or well
sub-bundling) protocol.openflow, hence lets embed in
protocol.openflow_netty protocol.openflow and lets override in the
class path what we need. For instance protocol.openflow embeds
openflowJ while we need openflow_netty, so making sure that
openflow_netty appears first in the classpath of the bundle ensure the
proper picking of the resources.
- In protocol.openflow_netty the portions that need to be
reimplemented are:
	# The activator, which largely is like the one of the
	protocol.openflow bundle with the enhancements of
	EnhacedController instantiation.
	# The logic netty specific, provide by Colin/Muthu
So the protocol.openflow_netty is just containing that logic and the
rest comes mint from protocol.openflow.
- The EnhancedActivator of protocol.openflow_netty extends the base
class being the protocol.openflow activator, and just override the
instantiation of EnhancedController when Controller.class was being
invoked.
- Remove openflowj_netty from distribution, is not needed because is
embedded in the protocol.openflow_netty
- Remove embedding in openflowj_netty of the netty dependency given
it's being used as a standalone bundle.

Change-Id: I27d7f6bb45ede88c077bead640cc69c74e49a9d3
Signed-off-by: Giovanni Meo <gmeo@cisco.com>