FRM event sync developement
- Created necessary logic for distributing entry programming among the
FRM instances in the cluster.
- Made serializable Status so it can be distributed.
- Hooked FRM with connection manager, such that the FRM programming
can be distributed if the node is not local
- In case of non-local node someone picks it and process it, then the
status is returned back to caller.
- At the moment the whole logic to timeout entries that don't get an
answer is missing, will come along.
- HostTracker ages out the ARP entries in clustered environment
because when refreshed the timeout was not being refreshed
clusterwide, probably need a better way to deal with the timing out
-Added license in files missing it.
- Fixed integration test failing, now in fact we can no longer use
fake values else the connection manager will not report them as
attached to local controller and will stuck the test case

Signed-off-by: Giovanni Meo <gmeo@cisco.com>
Change-Id: I74b7535371f61759e89883c66ba9b22596d7dddc