Prepare for mvn release:prepare release:perform operation
- Make sure every bundle include a proper scm section
- Provided a pom to be used for release operations in the root
directory, this is due to some implicit assumptions
maven-release-manager does when performing a release. The
maven-release-manager in fact calculate a tag name starting from the
developerConnection and removing from the URL layers based on how
deep the release pom is versus the assumed root. This is hardcoded in:
http://svn.apache.org/viewvc/maven/release/tags/maven-release-2.3.2/maven-release-manager/src/main/java/org/apache/maven/shared/release/util/ReleaseUtil.java?view=markup
Line 182 (this logic is not really specific to 2.3.2 version, but
pretty much in every version).
The tag calculation is fine if applied to SCM live SVN but not to GIT
hence this limitation of having to place the root pom in the root
itself.
- Removed uneeded distributions like sdk and parent, because are
combined in the main distribution.
- Corrected some wrong dependencies in clustering
- Added parent poms to distribution so to simplify the deploy of the
artifacts.
- Disable signing of openflowJ during release, it hangs
- Forcing maven-release-plugin to version 2.3.2, latest today (2.4.1)
has some issues.
- Updated README.OPENDAYLIGHT with process for releasing artifacts
- pom file in the root directory is merely a pointer to the main
distribution, this doesn't change any of the current workflow and
enable release support
- Create a profile triggered when the -DDOBUILDRELEASE is set which
exclude modules that are still depending on snapshots not under
controller project

Signed-off-by: Giovanni Meo <gmeo@cisco.com>
Change-Id: I7d18fd7a5efd5809d3fc77d4e0ddc302296d9d98