Added CorsFilter to enable secure cross site scripting
This is in addition to Ed's original Cors Filter changes. Default Cors Config
doesnt seem to work in certain scenarios. Added some custom configurations
and also added it per-bundle (started with Flow & i will add it to other bundles
once this is verified).

See: http://tomcat.apache.org/tomcat-7.0-doc/config/filter.html#CORS_Filter
And: http://en.wikipedia.org/wiki/Cross-origin_resource_sharing
This is done to allow a web page using javascript to be able to make calls
to our REST APIs even though it does not originate in our domain.

Added CorsFilter bundle in Third Party to bring in the class as a Fragment
on the org.apache.catalina bundle.

Added CorsFilter to the web/root web.xml file so it will be used for
all WebApps.

DO NOT MERGE YET.  NEEDS TO BE TRIED OUT BY THE FOLKS TRYING TO USE CORS.

Change-Id: I5fc6a53f2046816984fab722b841730c0eee396a
Signed-off-by: Madhu Venugopal <vmadhu@cisco.com>