Few additional fixes to enable Client script access to Northbound.
After Tomcat 7, httpOnly is true by default. This causes the httponly flag set on the Cookie.
This prevents the Scripts to access the cookie. setting it to false by default.
Due to the recent fix on disabling authentication for OPTIONS, if a script uses OPTIONS for an
NB-API, the username is null causing null pointer exception. Hence adding null pointer safety
checks on all the northbound APIs.

Change-Id: I726fd058b629c04695bf07d9fd52f6483d193b79
Signed-off-by: Madhu Venugopal <vmadhu@cisco.com>