Update patch set 1
Patch Set 1:

(1 comment)

I was trying to verify this fix, I see that now it doesn't fail like it failed before, however now it downloads guava version 12 and not 14 as specified in guava.version property, and from the following repository (please note - not a maven repo):
Downloading: http://nexus.opendaylight.org/content/repositories/ebr-bundles-release/com/google/guava/guava/12.0/guava-12.0.pom

Looks like this repository should be removed or reodered in the pom?
Also (of course not related directly to this patch), why the repositories are defined in so many pom files? It will be much more maintainable (and for other purposes as well) to have a root pom for the project that contains the repos in one central place.

Patch-set: 1
Reviewer: Gerrit User 739 <739@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0