Create change
Patch-set: 1
Change-id: Ie2a92ad99d35367a1d89dca390e7af54321bfe7c
Subject: Controller network interface enhancements: - Provide a configuration parameter "of.address = " in config.ini which allows a specific IP address to be specified.   If a valid IP address is specified, the controller will ONLY accept connections from switches with destination address   matching the specified configured address. If an address it NOT specified (which is the default case) or is syntactically   incorrect, the controller will use "any" (wildcard) address. The default case is what currently been implemented. - If the IP address is specified, the controller will determine the networking interface (eg, eth0) on which the   IP address is configured. If the interface is NOT up and is changed status between up and down, a log will be generated   to alert the network operation to take corrective actions. If the address is not specified (or is specified with an invalid   address syntax), the controller will use any interfaces and will NOT monitor the status of these interfaces.
Branch: refs/heads/master
Commit: 2e1c698a2bdfcbc3582fc72bf75424e1dc11f263