Bug #60 : JSON list responses from the NB-APIs returns native object (instead of List) when the API returns a single element.
We use Jersey and JAXB frameworks for the NB-API implementation.
Jersey by default uses Jettison for the JSON marshalling/demarshalling. It seems like there is a bug in Jettison for this case.
And the recommended approach is to use Jackson instead.
Among various approaches to replace Jettison with Jackson (including package scanning, etc..), the only easy approach in an
OSGi environment is to directly refer to the JerseyJaxbJsonProvider class from the Northbound application.
JerseyJaxbJsonProvider has dependancies on jersey-xc which is also pulled in.

Change-Id: I56a9c3b16324947abe15a9681671e2626a5de958
Signed-off-by: Madhu Venugopal <vmadhu@cisco.com>