Bug #60 : JSON list responses from the NB-APIs returns native object (instead of List) when the API returns a single element.
We use Jersey and JAXB frameworks for the NB-API implementation.
Jersey by default uses Jettison for the JSON marshalling/demarshalling. It seems like there is a bug in Jettison for this case.
And the recommended approach is to use Jackson instead.
Among various approaches to replace Jettison with Jackson (including package scanning, etc..), the only easy approach in an
OSGi environment is to directly refer to the JerseyJaxbJsonProvider class from the Northbound application.
JerseyJaxbJsonProvider has dependancies on jersey-xc which is also pulled in.

The Northbound Integration Test has inbuilt assumption that the NB-APIs use Jettison & hence a whole bunch of such bugs were
hidden and not visible to the tests as well. The correct approach is to use JAXB bindings in the IT so that immaterial of
Jettison or Jackson as the NB Marshalling infra, the IT need not change.
But, that is a bigger change and hence to restrict to the changes being addressed, this changeset just fixed these incorrect
assumptions and continued to use the Jettison library for demarshalling (while the actual Northbound API uses Jackson).

Also note that, such infra changes call for bumping both the NB-API version and the bundle version.
But, given the fact that the API freeze is scheduled later in October and more NB related changes are coming, the version bump
process can be scheduled later to ease the development work.

Change-Id: I47d8309dfbe11c251a30316bca37c7823cb0325a
Signed-off-by: Madhu Venugopal <vmadhu@cisco.com>