Make FRM to finally distribute the work orders
- Transformed frm.{workOrder|workStatus} caches to be transactional,
as a stop gap to an issue we have run into.
- The transformation of FRM caches in transactional opens an issue
where if used in a context of a transaction the FRM updates will not
travel in the cluster. We use the frm.{workOrder|workStatus} caches as
a way to synchronize operations hence this communications need to
happens also before the transaction commits. To avoid then this
chicken-and-egg problem we are forcing the distributeWorkOrder
business logic to execute off a different thread just to avoid to be
held by the transaction.
This is a stop gap solution till we caracterize and fix the
non-transactional cache issue seen

Signed-off-by: Giovanni Meo <gmeo@cisco.com>
Change-Id: I8e8dd3f38f037dc62252ea1e271e0972e32019f0