Update patch set 1
Patch Set 1:

(1 comment)

Alissa, wondering, given your are touching this area, if wouldn't be possible to centralize the properties for the integration tests in the commons.integrations artifact, and the one for the modules under test under common.opendaylight. To remove this repetition in all the places.

Patch-set: 1
Label: Code-Review=0