Adding Tri-states to the connection manager locality check.
As of now, the connection manager's isLocal() provides a binary true/false for a locality.
But is missing a key state of not-connected. This is now included as part of a new API (getLocalityStatus).
Retaining the isLocal as a deprecated method to be removed later once all the users have migrated.

Change-Id: I0feddf2a86770740f441060bd411f28d1bb15610
Signed-off-by: Madhu Venugopal <vmadhu@cisco.com>