Provide consistent hash code for enums and use it too
- Java enums if used as key, works well in a single JVM, but across
JVM the hashcode generated by them is not guaranteed to be the same
and that breaks the calculation of a cluster wide consistent hash
code, needed by the clustering services. The fix is to provide a
consistent hashcode for the Enum and to use it in the dependent data
structures.

Change-Id: Iee80f262253a287e8bc8be4d5371d0bd407f0462
Signed-off-by: Giovanni Meo <gmeo@cisco.com>