Update patch set 2
Patch Set 2: Code-Review-2

Alissa, this patch cannot go in, because moving in dependencyManagement the spring components, the distribution will not be able to pack into the plugins directory. The net effect is that if this patch goes in the build will go on fine, but the final distribution will fail to start the whole UI and northbound subsystems because the spring dependencies are not there at runtime.

Patch-set: 2
Reviewer: Gerrit User 19 <19@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-2