Reverting the FRM WorkOrder caches to NON_TRANSACTIONAL
After Giovanni's recent hashCode fix, most of the cache issue issues are resolved and it also renders itself to making
FRM WorkOrder back NON_TRANSACTIONAL. Also reverted the need to have a seperate thread to handle this processing.
Now it is back to normal and how it is supposed to be.

Change-Id: I0abbe66ea0ec7203ef306bc73d848112110f6460
Signed-off-by: Madhu Venugopal <vmadhu@cisco.com>