Removing the Container Aware dependency from Clustering Services.
Clustering Services implementation is ContainerAware to handle the case of Container deletion.
More appropriate way to handle it is to have an explicit API in global cluster manager to expose
a functionality to clean all the caches assocaited with a Container. This API can be used by
the Container Manager to clean up the caches.

Change-Id: I9758a01da13ffa7d951307a116df34a7fe9c19f6
Signed-off-by: Madhu Venugopal <vmadhu@cisco.com>