Update patch set 2
Patch Set 2:

Alissa, in order to sanitize the patch do this:
1) Build the controller
2) Run it
3) on the OSGi console do ss -s INSTALLED

you should not see any bundles listed in #3 instead with your patch you would see plenty of them because the dependencies at runtime are missing. Hope it helps.
I would love to see another submission for this, i myself i'm not familiar with dependencyManagement section of maven, so i hope Ed can give more details on it.

Patch-set: 2