Update patch set 1
Patch Set 1:

Madhu, 
I indeed see that the northbound IT test imports jettison classes.
So, just to make sure I got your comment right - jettison should stay in northbound IT module, but should be removed from the rest of the project? (if that change will work in current pom structure, of course)

Patch-set: 1