Update patch set 1
Patch Set 1:

Alissa,

This is a catch-22 problem here.
If we have to do it right, we have to remove jettison from Northbound IT as well. By removing it from IT dependencies, then the NorthboundIT.java wont compile as it uses Jettison client library extensively.
Instead we have to use JAXB/Jersey client auto-generated by enunciate.

Unfortunately, the JAXB/Jersey clients that the enunciate generates are not published and hence the Northbound IT must have add a local dependency to the client library that is built @ opendaylight/northbound/<NORTHBOUND-FUNCTIONAL-MODULE>/target/site/wsdocs/full-client.jar

I know this is more work for the work you started :-), But my 2C is to get this done sooner than later. Can you please look into that ?

Patch-set: 1