Update patch set 1
Patch Set 1:

Madhu, it sounds indeed like a bigger change than this patch especially since it involves code change and area which I am not familiar with.
I suggest to merge this patch because it does improve current state of things and makes easier maintenance.
Regardless of that, it will be good to see if another change that removes jettison as you suggested is possible.
What do you think?

Patch-set: 1