Update patch set 2
Patch Set 2:

Giovanni, I tried now to perform the steps you mentioned  on the latest master, without my patch.

However, when doing ss -s INSTALLED, it is not empty. 
The following 4 bundles are displayed:

17	INSTALLED   org.opendaylight.controller.model.flow-statistics_1.0.0.SNAPSHOT
78	INSTALLED   org.opendaylight.controller.model.flow-service_1.0.0.SNAPSHOT
81	INSTALLED   org.opendaylight.controller.sal-binding-broker-impl_1.0.0.SNAPSHOT
97	INSTALLED   org.opendaylight.controller.model.flow-base_1.0.0.SNAPSHOT


Is this OK? Perhaps the problem is not related to my change if this are the results without it?
Or I misunderstood something?

Patch-set: 2